% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_isotonic.R
\name{bayesian_isotonic}
\alias{bayesian_isotonic}
\title{Bayesian isotonic regression}
\usage{
bayesian_isotonic(
  data_grouped = NULL,
  stan_fit = NA,
  stan_path = NA,
  stan_args = list(local_dof_stan = 1, global_dof_stan = 1, alpha_scale_stan = 1),
  sample_from_prior_only = F,
  conf_level = 0.5,
  conf_level_direction = "both",
  sig_threshold = c(0.005, 0.01, 0.05),
  verbose = F,
  n_mc_warmup = 2500,
  n_mc_samps = 5000,
  mc_chains = 1,
  mc_thin = 1,
  mc_stepsize = 0.1,
  mc_adapt_delta = 0.99,
  mc_max_treedepth = 15,
  return_as_stan_object = F,
  tol = .Machine$double.eps^0.5,
  stan_seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{data_grouped}{a tibble or data.frame with named columns y and n indicating,
respectively, the number of outcomes and number of bernoulli trials. Each
row corresponds to a different category of the predictor. \strong{The rows are assumed
to be ordered properly, such that the first row corresponds to the first
category of the predictor, the second row the second category, and so forth}}
}
\description{
bayesian_isotonic
}
\examples{

}
\references{
\insertRef{boonstra2020b}{isotonicBayes}
}
