% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gaipv_prior.R
\name{generate_gaipv_prior}
\alias{generate_gaipv_prior}
\title{Generate draws from GAIPV prior}
\usage{
generate_gaipv_prior(
  n_cats_seq,
  shape_seq,
  n_draws = 10000,
  seed = sample(.Machine$integer.max, 1)
)
}
\arguments{
\item{n_cats_seq}{A vector of one or more positive integers giving the
desired values of K over which to evaluate}

\item{shape_seq}{A vector of one or more positive numbers giving the
set of shape parameters s over which to evaluate}

\item{n_draws}{a positive integer giving the number of random draws from the
t-distributions with which to estimate expectations}

\item{seed}{a positive integer giving the random seed, provided for reproducibility.}
}
\value{
a tibble with number of rows equal to the product of the lengths of
\code{n_cats_seq} and \code{shape_seq} and three columns: the current number of categories,
the current shape, and the calculated effective number of prior parameters
}
\description{
Generate draws from the GAIPV prior and then calculate the
effective number of parameters according to this prior.
}
\examples{
generate_gaipv_prior(n_cats_seq = c(1, 5, 10), shape_seq = c(1/400, 400), n_draws = 1e4)


}
\references{
\insertRef{boonstra2020b}{isotonicBayes}

\insertRef{piironen2017hyperprior}{isotonicBayes}
}
