% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hsipv_prior.R
\name{generate_hsipv_prior}
\alias{generate_hsipv_prior}
\title{Generate draws from HSIPV prior}
\usage{
generate_hsipv_prior(
  n_cats_seq,
  scale_seq,
  local_dof = 1,
  global_dof = 1,
  slab_precision = 1,
  n_draws,
  seed = sample(.Machine$integer.max, 1)
)
}
\arguments{
\item{n_cats_seq}{A vector of one or more positive integers giving the
desired values of K over which to evaluate}

\item{scale_seq}{A vector of one or more positive numbers giving the
set of scale parameters c over which to evaluate}

\item{local_dof}{a positive number giving the degrees of freedom for the
t-distribution of the local shrinkage parameter.
local_dof = global_dof = 1 corresponds to the horseshoe}

\item{global_dof}{a positive number giving the degrees of freedom for the
t-distribution of the global shrinkage parameter.
local_dof = global_dof = 1 corresponds to the horseshoe}

\item{slab_precision}{a positive number giving the precision of the slab
parameter for the regularized horseshoe. Defaults to 1, which is the only
value considered in the manuscript}

\item{n_draws}{a positive integer giving the number of random draws from the
t-distributions with which to estimate expectations}

\item{seed}{a positive integer giving the random seed, provided for reproducibility.}
}
\value{
a tibble with number of rows equal to the product of the lengths of
\code{n_cats_seq} and \code{scale_seq} and three columns: the current number of categories,
the current scale, and the calculated effective number of prior parameters
}
\description{
Generate draws from the HSIPV prior and then calculate the
effective number of parameters according to this prior.
}
\examples{
generate_hsipv_prior(n_cats_seq = c(1, 5, 10), scale_seq = c(1/400, 400), n_draws = 1e4)


}
\references{
\insertRef{boonstra2020b}{isotonicBayes}

\insertRef{piironen2017hyperprior}{isotonicBayes}
}
