% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotonicBayes-package.R
\docType{package}
\name{isotonicBayes-package}
\alias{isotonicBayes-package}
\alias{isotonicBayes}
\title{The 'isotonicBayes' package.}
\description{
This R package implements the isotonic probability vector priors
developed in Boonstra, Owen, and Kang (2020). The main function is
\code{bayesian_isotonic}, through which users can fit isotonic regression models
with a binary outcome regressed against a categorical predictor.
}
\references{
\insertRef{boonstra2020b}{isotonicBayes}

Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
