% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grouped_data.R
\name{make_grouped_data}
\alias{make_grouped_data}
\title{Categorize continuous data}
\usage{
make_grouped_data(x, y, breaks = NULL)
}
\arguments{
\item{x}{A vector of continuous predictors}

\item{y}{A vector of Bernoulli outcomes with length equal to length(x)}

\item{breaks}{A vector of values used to categorize x}
}
\description{
This is a helper function to turn a vector of continuous predictors x
and corresponding vector of binary outcomes y into a data.frame of a form
expected by the function \code{\link[=bayesian_isotonic]{bayesian_isotonic()}}
}
\references{
\insertRef{boonstra2020b}{isotonicBayes}
}
