% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_hs_scale.R
\name{solve_for_hs_scale}
\alias{solve_for_hs_scale}
\title{solve_hs_scale}
\usage{
solve_for_hs_scale(
  target_mean,
  local_dof = 1,
  global_dof = 1,
  slab_precision = 1,
  n,
  sigma = 2,
  tol = .Machine$double.eps,
  max_iter = 100,
  n_sim = 1e+06
)
}
\arguments{
\item{target_mean}{}
}
\description{
solve_hs_scale
}
\examples{

}
\references{
\insertRef{boonstra2020}{isotonicBayes}
}
