% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_for_hs_scale.R
\name{solve_for_hs_scale}
\alias{solve_for_hs_scale}
\title{solve_for_hs_scale}
\usage{
solve_for_hs_scale(
  target_mean,
  local_dof = 1,
  global_dof = 1,
  slab_precision = 1,
  n,
  sigma = 2,
  tol = .Machine$double.eps,
  max_iter = 100,
  n_sim = 1e+06
)
}
\arguments{
\item{target_mean}{a number in 0,1 giving the desired proportion of non-zero
parameters to which the returned scale value should correspond.}

\item{local_dof}{a positive number giving the degrees of freedom for the
t-distribution of the local shrinkage parameter.
local_dof = global_dof = 1 corresponds to the horseshoe}

\item{global_dof}{a positive number giving the degrees of freedom for the
t-distribution of the global shrinkage parameter.
local_dof = global_dof = 1 corresponds to the horseshoe}

\item{slab_precision}{a positive number giving the precision of the slab
component of the regularized horseshoe.}

\item{n}{a positive integer giving the sample size}

\item{sigma}{the presumed standard deviation parameter for the measurement error.
If the outcome is binary, then sigma=2 corresponds to the most conservative
choice in the sense of yielding more shrinkage than desired \insertCite{piironen2017hyperprior}{isotonicBayes}}

\item{tol}{a small number giving the acceptable numerical tolerance
for determining equality}

\item{max_iter}{a positive integer giving the maximum number of iterations
to proceed without convergence before giving up}

\item{n_sim}{a positive integer giving the number of random draws from the
t-distributions with which to estimate expectations}
}
\description{
solve_for_hs_scale
}
\examples{
solve_for_hs_scale(target_mean = 0.1, n = 100, n_sim = 1e5)


}
\references{
\insertRef{boonstra2020b}{isotonicBayes}

\insertRef{piironen2017hyperprior}{isotonicBayes}
}
